FGP-1031: refactored create-offer to use strategy pattern
